refactor pm quality fairness preview router
