+++ src/app/router_registry.py
@@ -153,6 +153,9 @@ from app.routers.dpm_command_center_pm_quality_fairness_detail import (
 from app.routers.dpm_command_center_pm_quality_fairness_lookup import (
     router as dpm_command_center_pm_quality_fairness_lookup_router,
 )
+from app.routers.dpm_command_center_pm_quality_fairness_preview import (
+    router as dpm_command_center_pm_quality_fairness_preview_router,
+)
 from app.routers.dpm_command_center_pm_quality_policies import (
     router as dpm_command_center_pm_quality_policies_router,
 )
@@ -533,6 +536,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_wave_outcome_reviews_router)
     app.include_router(dpm_command_center_pm_quality_router)
     app.include_router(dpm_command_center_pm_quality_ai_router)
+    app.include_router(dpm_command_center_pm_quality_fairness_preview_router)
     app.include_router(dpm_command_center_pm_quality_fairness_router)
     app.include_router(dpm_command_center_pm_quality_fairness_detail_router)
     app.include_router(dpm_command_center_pm_quality_fairness_lookup_router)
